chore(modelarmor): split base and runnables modules. added missing test cases.
